+++ server/resources/reserva.py
@@ -1,7 +1,7 @@
 from flask import request as req
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required, get_jwt
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from sqlalchemy.exc import SQLAlchemyError
 
 from datetime import datetime
@@ -16,15 +16,19 @@ blp = Blueprint("Reservas", __name__, description="Operações com reservas.")
 
 @blp.route('/reservas')
 class ReservaList(MethodView):
+  @jwt_required()
   @blp.response(200, ReservaSchema(many=True))
   def get(self):
-    reservas = ReservaModel.query.all()
+    usuario_id = get_jwt_identity()
+
+    reservas = ReservaModel.query.filter(ReservaModel.id_usuario == usuario_id).all()
 
     return reservas
   
 
 @blp.route('/reserva')
 class Reserva(MethodView):
+  @jwt_required()
   @blp.response(200, ReservaSchema)
   def get(self):
     reserva_id = req.args.get('id')
@@ -36,6 +40,8 @@ class Reserva(MethodView):
   @blp.arguments(ReservaSchema)
   @blp.response(201, ReservaSchema)
   def post(self, reserva_data):
+    usuario_id = get_jwt_identity()
+    reserva_data["id_usuario"] = usuario_id
     reserva = ReservaModel(**reserva_data)
 
     try:
@@ -50,39 +56,47 @@ class Reserva(MethodView):
   @blp.arguments(ReservaSchema)
   @blp.response(200, ReservaSchema)
   def put(self, reserva_data):
+    usuario_id = get_jwt_identity()
+
     reserva_id = req.args.get('id')
     reserva = ReservaModel.query.get(reserva_id)
 
-    if reserva:
-      reserva.id_viagem = reserva_data["id_viagem"]
-      reserva.id_usuario = reserva_data["id_usuario"]
-      reserva.assento = reserva_data["assento"]
-      reserva.forma_pagamento = reserva_data["forma_pagamento"]
-      reserva.atualizado_em = datetime.now().isoformat()
-    else:
-      reserva = ReservaModel(id=reserva_id, **reserva_data)
+    if reserva.id_usuario == usuario_id:
+      if reserva:
+        reserva.id_viagem = reserva_data["id_viagem"]
+        reserva.id_usuario = usuario_id
+        reserva.assento = reserva_data["assento"]
+        reserva.forma_pagamento = reserva_data["forma_pagamento"]
+        reserva.atualizado_em = datetime.utcnow()
+      else:
+        reserva = ReservaModel(id=reserva_id, **reserva_data)
 
-    try:
-      db.session.add(reserva)
-      db.session.commit()
-    except SQLAlchemyError:
-      abort(500, "An error ocurred while updating item from table 'reserva'.")
+      try:
+        db.session.add(reserva)
+        db.session.commit()
+      except SQLAlchemyError:
+        abort(500, "An error ocurred while updating item from table 'reserva'.")
 
-    return reserva
+      return reserva
+    abort(401, "Identitification error.")
 
   @jwt_required()
   @blp.response(204)
   def delete(self):
+    usuario_id = get_jwt_identity()
+
     reserva_id = req.args.get('id')
     reserva = ReservaModel.query.get(reserva_id)
 
-    if reserva:
-      try:
-        db.session.delete(reserva)
-        db.session.commit()
-      except SQLAlchemyError:
-        abort(500, "An error ocurred while deleting item from table 'reserva'.")
-    else:
-      abort(404, "Item not found.")
+    if reserva.id_usuario == usuario_id:
+      if reserva:
+        try:
+          db.session.delete(reserva)
+          db.session.commit()
+        except SQLAlchemyError:
+          abort(500, "An error ocurred while deleting item from table 'reserva'.")
+      else:
+        abort(404, "Item not found.")
 
-    return {"message": "Reserva excluida."}
+      return {"message": "Reserva excluida."}
+    abort(401, "Identification error.")
