+++ server/resources/sentido.py
@@ -2,6 +2,7 @@ from flask import request as req
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 
 from datetime import datetime
 
@@ -35,47 +36,63 @@ class Sentido(MethodView):
 
     return sentido
 
+  @jwt_required()
   @blp.arguments(SentidoSchema)
   @blp.response(201, SentidoSchema)
   def post(self, sentido_data):
-    sentido = SentidoModel(**sentido_data)
+    usuario_admin = get_jwt()["admin"]
+    
+    if usuario_admin:
+      sentido = SentidoModel(**sentido_data)
 
-    try:
-      db.session.add(sentido)
-      db.session.commit()
-    except SQLAlchemyError:
-      abort(500, "An error ocurred while inserting item to table 'sentido'.")
+      try:
+        db.session.add(sentido)
+        db.session.commit()
+      except SQLAlchemyError:
+        abort(500, "An error ocurred while inserting item to table 'sentido'.")
 
-    return sentido
+      return sentido
+    
+    abort(401, "Unauthorized access.")
   
+  @jwt_required()
   @blp.arguments(SentidoSchema)
   @blp.response(200, SentidoSchema)
   def put(self, sentido_data):
-    sentido_id = req.args.get('id')
-    sentido = SentidoModel.query.get(sentido_id)
+    usuario_admin = get_jwt()["admin"]
+    
+    if usuario_admin:
+      sentido_id = req.args.get('id')
+      sentido = SentidoModel.query.get(sentido_id)
 
-    if sentido:
-      sentido.id_linha = sentido_data["id_linha"]
-      sentido.sentido = sentido_data["sentido"]
-      sentido.ponto_partida = sentido_data["ponto_partida"]
-      sentido.ponto_destino = sentido_data["ponto_destino"]
-      sentido.horario_inicio = sentido_data["horario_inicio"]
-      sentido.horario_fim = sentido_data["horario_fim"]
-      sentido.atualizado_em = datetime.now().isoformat()
-    else:
-      sentido = SentidoModel(id=sentido_id, **sentido_data)
+      if sentido:
+        sentido.id_linha = sentido_data["id_linha"]
+        sentido.sentido = sentido_data["sentido"]
+        sentido.ponto_partida = sentido_data["ponto_partida"]
+        sentido.ponto_destino = sentido_data["ponto_destino"]
+        sentido.horario_inicio = sentido_data["horario_inicio"]
+        sentido.horario_fim = sentido_data["horario_fim"]
+        sentido.atualizado_em = datetime.now().isoformat()
+      else:
+        sentido = SentidoModel(id=sentido_id, **sentido_data)
 
-    return sentido
+      return sentido
+    abort(401, "Unauthorized access.")
   
+  @jwt_required()
   @blp.response(204)
   def delete(self):
-    sentido_id = req.args.get('id')
-    sentido = SentidoModel.query.get(sentido_id)
+    usuario_admin = get_jwt()["admin"]
+    
+    if usuario_admin:
+      sentido_id = req.args.get('id')
+      sentido = SentidoModel.query.get(sentido_id)
 
-    if sentido:
-      db.session.delete(sentido)
-      db.session.commit()
-    else:
-      abort(404, "Item not found.")
+      if sentido:
+        db.session.delete(sentido)
+        db.session.commit()
+      else:
+        abort(404, "Item not found.")
 
-    return {"message", "Sentido excluido."}
+      return {"message", "Sentido excluido."}
+    abort(401, "Unauthorized access.")
